+++ src/GafferTestModule/GafferTestModule.cpp
@@ -70,6 +70,11 @@ static boost::python::tuple countContextHash32CollisionsWrapper( int entries, in
 	return boost::python::make_tuple( std::get<0>(result), std::get<1>(result), std::get<2>(result), std::get<3>(result) );
 }
 
+static float asFloat32( const float value )
+{
+	return value;
+}
+
 BOOST_PYTHON_MODULE( _GafferTest )
 {
 
@@ -82,7 +87,7 @@ BOOST_PYTHON_MODULE( _GafferTest )
 			)
 		);
 
-
+	def( "asFloat32", &asFloat32 );
 	def( "testRecursiveChildIterator", &testRecursiveChildIterator );
 	def( "testFilteredRecursiveChildIterator", &testFilteredRecursiveChildIterator );
 	def( "testMetadataThreading", &testMetadataThreadingWrapper );